server/pyproject.toml
[project]
name = "ecommerce-agent"
version = "0.0.1"
description = "Ecommerce agent"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.32.1",
    "pydantic-settings>=2.0",
    "google-genai>=1.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = ["pytest", "httpx", "ruff"]

[tool.ruff]
line-length = 100